Rename ipa_helper to ec2_ipa_helper.
- Pass ssh_user to base job class.
- Move temp key pair handling to a context manager.
