+++ mash/utils/ec2.py
@@ -18,6 +18,10 @@
 
 import boto3
 
+from contextlib import contextmanager, suppress
+
+from mash.utils.mash_utils import generate_name, get_key_from_file
+
 
 def get_client(service_name, access_key_id, secret_access_key, region_name):
     """
@@ -29,3 +33,27 @@ def get_client(service_name, access_key_id, secret_access_key, region_name):
         aws_secret_access_key=secret_access_key,
         region_name=region_name,
     )
+
+
+@contextmanager
+def temp_key_pair(
+    access_key_id, region, secret_access_key, ssh_private_key_file
+):
+    """
+    Create a temp key pair in the account and region.
+
+    Return the key pair name.
+    """
+    try:
+        key_name = generate_name()
+        client = get_client(
+            'ec2', access_key_id, secret_access_key, region
+        )
+        ssh_public_key = get_key_from_file(ssh_private_key_file + '.pub')
+        client.import_key_pair(
+            KeyName=key_name, PublicKeyMaterial=ssh_public_key
+        )
+        yield key_name
+    finally:
+        with suppress(Exception):
+            client.delete_key_pair(KeyName=key_name)
